[REF] Move all the generic functions to the parent
This is for clarity, note that I need agreement on
https://lab.civicrm.org/dev/core/-/issues/2745
to decide whether to keep 'getExposedFields' or not - hence
that is not moved at this stage
